# api/openapi.yaml
openapi: 3.0.3
info:
  title: Cockpit Management API
  version: 7.0.0
  description: Resource-oriented management API for Cockpit.
servers:
  - url: /api
    description: Local management server
components:
  schemas:
    StatusOk:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    AuthFile:
      type: object
      properties:
        id:
          type: string
        auth_index:
          type: string
        name:
          type: string
        type:
          type: string
        provider:
          type: string
        label:
          type: string
        email:
          type: string
        status:
          type: string
        status_message:
          type: string
        disabled:
          type: boolean
        unavailable:
          type: boolean
        runtime_only:
          type: boolean
        size:
          type: integer
        modtime:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_refresh:
          type: string
          format: date-time
        priority:
          type: integer
        usage:
          type: object
          additionalProperties: true
        usage_available:
          type: boolean
    AuthFilesEnvelope:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuthFile'
    CreateAuthFileRequest:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
        content:
          type: object
          additionalProperties: true
    AuthFileCreateResponse:
      type: object
      required: [status, name]
      properties:
        status:
          type: string
          example: ok
        name:
          type: string
    AuthFilePatchRequest:
      type: object
      properties:
        disabled:
          type: boolean
        priority:
          type: integer
          nullable: true
    AuthFileDeleteResponse:
      type: object
      required: [status, name]
      properties:
        status:
          type: string
          example: ok
        name:
          type: string
...
paths:
...
  /auth-files:
    get:
      summary: List managed auth files
      responses:
        '200':
          description: Auth file collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthFilesEnvelope'
    post:
      summary: Create or upload an auth file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthFileRequest'
      responses:
        '200':
          description: Auth file created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthFileCreateResponse'
        '400':
          description: Invalid auth file payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth-files/{name}:
    patch:
      summary: Update editable auth file fields
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthFilePatchRequest'
      responses:
        '200':
          description: Auth file updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusOk'
        '400':
          description: Invalid auth file mutation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Auth file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete an auth file
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Auth file deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthFileDeleteResponse'
        '404':
          description: Auth file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...